pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentscope-adaptive-orchestrator"
version = "0.1.0"
description = "Adaptive orchestrator agent skeleton built on AgentScope-first principles"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "httpx>=0.27",
]

[project.optional-dependencies]
agentscope = [
  "agentscope>=0.1.0",
]

dev = [
  "pytest>=8.0",
]

[project.scripts]
adaptive-orchestrator = "main:cli"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
